# Server/rest-api-module/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: KIV/PIA REST API
  description: blah
  version: 0.0.1
servers:
  - url: http://localhost:8080/
    description: Local development server
...
paths:
...
  /projects:
    post:
      tags:
        - Projects
      summary: Creates a new project using the current user as the customer.
      description: >
        Creates a new translation project with the provided target language and content.
        The authenticated user is set as the customer for the project. (Only CUSTOMER can access this endpoint.)
      operationId: createProject
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
        required: true
      responses:
        '201':
          description: Project was successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        '400':
          description: Request body is not formatted correctly.
        '403':
          description: Current user is not authorized to create projects.
        '405':
          description: Unsupported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
    get:
      tags:
        - Projects
      summary: List all projects with optional filters.
      description: >
        Retrieves a list of all projects in the system, with optional filtering
        by state, target language, and feedback presence. (ALL ROLES USERS CAN ACCESS)
        Every user sees only projects they are authorized to view.
      operationId: listAllProjects
      parameters:
        - in: query
          name: state
          schema:
            $ref: '#/components/schemas/ProjectState'
          description: Filter projects by their state.
          required: false
        - in: query
          name: languageCode
          schema:
            type: string
            minLength: 2
            maxLength: 2
            description: Filter projects by target language (ISO 639-1 code).
          required: false
        - in: query
          name: hasFeedback
          schema:
            type: boolean
            description: Filter projects based on whether they have feedback.
          required: false


      responses:
        '200':
          description: Projects retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProjectsResponse"
        '403':
          description: Current user is not authorized to list projects.
        '405':
          description: Unsupported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: Get project by ID
      description: >
        Retrieves the details of a project by its unique ID.
        (ALL ROLES USERS CAN ACCESS this endpoint; user sees only projects they are authorized to view.)
      operationId: getProjectDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized - no valid authentication token provided.
        '403':
          description: Forbidden - insufficient permissions to access this resource.
        '404':
          description: Project not found
        '500':
          description: Server error.
  /projects/{id}/original:
    get:
      tags:
        - Projects
      summary: Download the original content of a project
      description: >
        Retrieves the original content file of the specified project.
        (customer  or assigned translator can perform this)
      operationId: downloadOriginalContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Original content retrieved successfully.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized - no valid authentication token provided.
        '403':
          description: Forbidden - insufficient permissions to access this resource.
        '404':
          description: Project not found.
        '500':
          description: Server error.
  /projects/{id}/translated:
    get:
      tags:
        - Projects
      summary: Download the translated content of a project (customer, admin, or assigned translator can perform this)
      description: >
        Retrieves the translated content file of the specified project.
        (customer or assigned translator can perform this)
      operationId: downloadTranslatedContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Translated content retrieved successfully.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized - no valid authentication token provided.
        '403':
          description: Forbidden - insufficient permissions to access this resource.
        '404':
          description: Project not found.
        '500':
          description: Server error.
    put:
      tags:
        - ProjectsWorkflow
      summary: Upload translated content for a project (Only translator for this project can perform this )
      description: >
        Allows the assigned translator to upload the translated content file
        for the specified project. (Only translator for this project can perform this )
      operationId: uploadTranslatedContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: Translated content uploaded successfully.
        '400':
          description: Invalid input data.
        '401':
          description: Unauthorized - no valid authentication token provided.
        '403':
          description: Forbidden - insufficient permissions to modify this resource.
        '404':
          description: Project not found.
        '500':
          description: Server error.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        name:
          type: string
          example: "John Doe"
        emailAddress:
          type: string
          format: email
          example: "john.doe@example.com"
        role:
          $ref: '#/components/schemas/UserRole'
        languages:
          type: array
          items:
            type: string
            example: "en"
          description: ISO language codes
        createdAt:
          type: string
          format: date-time
          example: "2025-10-15T09:33:20Z"
      required:
        - id
        - name
        - emailAddress
        - role
        - createdAt
    UserRole:
      type: string
      enum:
        - CUSTOMER
        - TRANSLATOR
        - ADMINISTRATOR
    
    #??? MAYBE HEADERS ???
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the created project
        targetLanguage:
          type: string
          example: "de"
        originalFileName:
          type: string
          example: "document.txt"
        translatedFileName:
          type: string
          example: "document_de.txt"
        customer:
          $ref:
            '#/components/schemas/User'
        translator:
          $ref:
            '#/components/schemas/User'
        state:
          $ref: '#/components/schemas/ProjectState'
        createdAt:
          type: string
          format: date-time
    ProjectState:
      type: string
      enum:
        - CREATED
        - ASSIGNED
        - COMPLETED
        - APPROVED
        - CLOSED

    ### PROJECT REQUESTS ###
    CreateProjectRequest:
      type: object
      required: [ languageCode, content ]
      additionalProperties: false
      properties:
        languageCode:
          type: string
          minLength: 2
          maxLength: 2
          example: "de"
          description: Target language ISO 639-1 code.
        content:
          type: string
          format: binary
          description: Text content to be translated.

    ### PROJECT RESPONSES ###
    ListProjectsResponse:
      type: array
      items:
        $ref: '#/components/schemas/Project'


    #------ Project Feedback Schemas -----
...